SWP-505 - update article instead of removing it on content push
