+++ src/SWP/Bundle/ContentBundle/Factory/AbstractArticleFactory.php
@@ -16,36 +16,22 @@ declare(strict_types=1);
 
 namespace SWP\Bundle\ContentBundle\Factory;
 
+use SWP\Bundle\ContentBundle\Hydrator\ArticleHydratorInterface;
 use SWP\Bundle\ContentBundle\Model\ArticleInterface;
-use SWP\Bundle\ContentBundle\Provider\RouteProviderInterface;
-use SWP\Component\Bridge\Model\ItemInterface;
 use SWP\Component\Bridge\Model\PackageInterface;
 
 abstract class AbstractArticleFactory implements ArticleFactoryInterface
 {
-    /**
-     * @var array
-     */
-    private $allowedTypes = [
-        ItemInterface::TYPE_PICTURE,
-        ItemInterface::TYPE_FILE,
-        ItemInterface::TYPE_TEXT,
-        ItemInterface::TYPE_COMPOSITE,
-    ];
-
-    /**
-     * @var RouteProviderInterface
-     */
-    protected $routeProvider;
+    protected $articleHydrator;
 
     /**
      * AbstractArticleFactory constructor.
      *
-     * @param RouteProviderInterface $routeProvider
+     * @param ArticleHydratorInterface $articleHydrator
      */
-    public function __construct(RouteProviderInterface $routeProvider)
+    public function __construct(ArticleHydratorInterface $articleHydrator)
     {
-        $this->routeProvider = $routeProvider;
+        $this->articleHydrator = $articleHydrator;
     }
 
     /**
@@ -58,62 +44,6 @@ abstract class AbstractArticleFactory implements ArticleFactoryInterface
         /** @var ArticleInterface $article */
         $article = $this->create();
 
-        $article->setBody($this->populateBody($package));
-        $article->setTitle($package->getHeadline());
-        if (null !== $package->getSlugline()) {
-            $article->setSlug($package->getSlugline());
-        }
-
-        $article->setLocale($package->getLanguage());
-        $article->setLead($this->populateLead($package));
-        $article->setMetadata($package->getMetadata());
-        // assign default route
-        $article->setRoute($this->routeProvider->getRouteForArticle($article));
-
-        return $article;
-    }
-
-    /**
-     * @param PackageInterface $package
-     *
-     * @return string
-     */
-    protected function populateLead(PackageInterface $package)
-    {
-        return trim($package->getDescription().implode('', array_map(function (ItemInterface $item) {
-            $this->ensureTypeIsAllowed($item->getType());
-
-            return ' '.$item->getDescription();
-        }, $package->getItems()->toArray())));
-    }
-
-    /**
-     * @param PackageInterface $package
-     *
-     * @return string
-     */
-    protected function populateBody(PackageInterface $package)
-    {
-        return $package->getBody().' '.implode('', array_map(function (ItemInterface $item) {
-            $this->ensureTypeIsAllowed($item->getType());
-
-            return $item->getBody();
-        }, $package->getItems()->toArray()));
-    }
-
-    /**
-     * @param string $type
-     *
-     * @throws \InvalidArgumentException
-     */
-    protected function ensureTypeIsAllowed(string $type)
-    {
-        if (!in_array($type, $this->allowedTypes)) {
-            throw new \InvalidArgumentException(sprintf(
-                'Item type "%s" is not supported. Supported types are: %s',
-                $type,
-                implode(', ', $this->allowedTypes)
-            ));
-        }
+        return $this->articleHydrator->hydrate($article, $package);
     }
 }
